test(apps/github-issues): move specs
